re-arrange files for seperate Phase and HearingFormStep5 components
